pyproject.toml
[project]
name = "telegram-ai-org"
version = "0.1.0"
description = "AI organization on Telegram — PM bot coordinates AI teams with shared context"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "pydantic>=2.0",
    "aiosqlite>=0.19",
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
